+++ CHANGELOG.md
@@ -9,7 +9,12 @@
 - added arrayFilter Class
 - added hook181 for auto finalize
 - added field for error reporting address in merchant backend
+
+####Changed
 - disabled Santander invoice in info.xml
+- replaced heidelpay/php-api with heidelpay/php-payment-api 
+- replaced heidelpay/php-customer-messages with eidelpay/php-message-code-mapper
+- changed namespaces according to new packages 
 
 ## v112